+++ lib/src/numdart/numbers/numbers.dart
@@ -4,6 +4,7 @@ export './complex_operations/complex_abs.dart';
 export './complex_operations/complex_conjugate.dart';
 export './complex_operations/complex_cos.dart';
 export './complex_operations/complex_divide_scalar.dart';
+export './complex_operations/complex_exp.dart';
 export './complex_operations/complex_multiply_scalar.dart';
 export './complex_operations/complex_truncate.dart';
 export 'complex.dart';